MB-61855 Handle IPv6 when converting to otp name
When searching for a node given by the user we would compare the
hostnames from /pools/default and a variety of ports with the user
string by using a string of the form '<hostname>:<port>'. This does not
work with IPv6 where the address needs to be in square brackets.

Testing
-------
1. Ran `recovery` with an IPv6 address on a cluster that was in
   IPv6-only mode. Node was found and the correct otp address was sent
   to ns_server.

Change-Id: Ic6574b7e946a9410d0e9736b699dec890977ff86
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/209830
Reviewed-by: Nour Elmenyawi <nour.elmenyawi@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Matt Hall <matt.hall@couchbase.com>