MB-64136 Better default paths + overrides
Previously, the default paths for the various directories were found
relative to the location of couchbase-cli. This was acceptable since
the commands that relied on this were only supposed to be run from the
CB Server location of couchbase-cli. With the new admin-tools, this is
no longer necessarily true.

This change now uses the relative path if available, but if not, we
fall back to the default paths. There are also documented override
flags which can be used in case of a non-default installation of CB
Server.

Testing
_______

Tested with default + non-default install + using override flags on
Linux, Mac and Windows.

Change-Id: I48e2aa44458fc1fa71265651f1d363320f22981c
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/219416
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Hall <matt.hall@couchbase.com>
Well-Formed: Restriction Checker