Add --eval param for server-eshell command
It makes is possible to test 99% of server-eshell command by
automated tests. The only thing that can't be tested is the start of
the shell process itself, but this part is unlikely to break.
The most common reason for the eshell command to break is when the
node started by the CLI can't establish a distribution connection to
the server node. Since --eval param executes code on the server node
(via rpc:call), it makes sure that the distribution connection is up.

Change-Id: I98514b4b4acc2348247b2bdc424f175a4aa015fd
Reviewed-on: https://review.couchbase.org/c/couchbase-cli/+/221409
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Hall <matt.hall@couchbase.com>